Cleanup: move unneeded classes to sub-package
